+++ packages/transactions/src/constants.ts
@@ -17,6 +17,7 @@ const UNCOMPRESSED_PUBKEY_LENGTH_BYTES = 64;
 const MEMO_MAX_LENGTH_BYTES = 34;
 const DEFAULT_CORE_NODE_API_URL = 'https://stacks-node-api.mainnet.stacks.co';
 
+// todo: add explicit enum values
 enum StacksMessageType {
   Address,
   Principal,
@@ -41,9 +42,16 @@ export function whenMessageType(messageType: StacksMessageType) {
 enum PayloadType {
   TokenTransfer = 0x00,
   SmartContract = 0x01,
+  VersionedSmartContract = 0x06,
   ContractCall = 0x02,
   PoisonMicroblock = 0x03,
   Coinbase = 0x04,
+  CoinbaseToAltRecipient = 0x05,
+}
+
+enum ClarityVersion {
+  Clarity1 = 1,
+  Clarity2 = 2,
 }
 
 /**
@@ -178,6 +186,7 @@ export {
   ChainID,
   StacksMessageType,
   PayloadType,
+  ClarityVersion,
   AnchorMode,
   TransactionVersion,
   PostConditionMode,