+++ packages/transactions/src/payload.ts
@@ -1,28 +1,45 @@
 import { concatArray, IntegerType, intToBigInt, intToBytes, writeUInt32BE } from '@stacks/common';
-import { COINBASE_BYTES_LENGTH, PayloadType, StacksMessageType } from './constants';
+import { ClarityVersion, COINBASE_BYTES_LENGTH, PayloadType, StacksMessageType } from './constants';
 
+import { BytesReader } from './bytesReader';
+import { ClarityValue, deserializeCV, serializeCV } from './clarity/';
+import { PrincipalCV, principalCV } from './clarity/types/principalCV';
+import { Address } from './common';
+import { createAddress, createLPString, LengthPrefixedString } from './postcondition-types';
 import {
-  MemoString,
+  codeBodyString,
   createMemoString,
-  serializeStacksMessage,
   deserializeAddress,
   deserializeLPString,
   deserializeMemoString,
-  codeBodyString,
+  MemoString,
+  serializeStacksMessage,
 } from './types';
-import { createAddress, LengthPrefixedString, createLPString } from './postcondition-types';
-import { Address } from './common';
-import { ClarityValue, serializeCV, deserializeCV } from './clarity/';
-
-import { BytesReader } from './bytesReader';
-import { PrincipalCV, principalCV } from './clarity/types/principalCV';
 
 export type Payload =
   | TokenTransferPayload
   | ContractCallPayload
   | SmartContractPayload
+  | VersionedSmartContractPayload
   | PoisonPayload
-  | CoinbasePayload;
+  | CoinbasePayload
+  | CoinbasePayloadToAltRecipient;
+
+export function isTokenTransferPayload(p: Payload): p is TokenTransferPayload {
+  return p.payloadType === PayloadType.TokenTransfer;
+}
+export function isContractCallPayload(p: Payload): p is ContractCallPayload {
+  return p.payloadType === PayloadType.ContractCall;
+}
+export function isSmartContractPayload(p: Payload): p is SmartContractPayload {
+  return p.payloadType === PayloadType.SmartContract;
+}
+export function isPoisonPayload(p: Payload): p is PoisonPayload {
+  return p.payloadType === PayloadType.PoisonMicroblock;
+}
+export function isCoinbasePayload(p: Payload): p is CoinbasePayload {
+  return p.payloadType === PayloadType.Coinbase;
+}
 
 export interface TokenTransferPayload {
   readonly type: StacksMessageType.Payload;
@@ -36,8 +53,10 @@ export type PayloadInput =
   | (TokenTransferPayload | (Omit<TokenTransferPayload, 'amount'> & { amount: IntegerType }))
   | ContractCallPayload
   | SmartContractPayload
+  | VersionedSmartContractPayload
   | PoisonPayload
-  | CoinbasePayload;
+  | CoinbasePayload
+  | CoinbasePayloadToAltRecipient;
 
 export function createTokenTransferPayload(
   recipient: string | PrincipalCV,
@@ -102,10 +121,19 @@ export interface SmartContractPayload {
   readonly codeBody: LengthPrefixedString;
 }
 
+export interface VersionedSmartContractPayload {
+  readonly type: StacksMessageType.Payload;
+  readonly payloadType: PayloadType.VersionedSmartContract;
+  readonly clarityVersion: ClarityVersion;
+  readonly contractName: LengthPrefixedString;
+  readonly codeBody: LengthPrefixedString;
+}
+
 export function createSmartContractPayload(
   contractName: string | LengthPrefixedString,
-  codeBody: string | LengthPrefixedString
-): SmartContractPayload {
+  codeBody: string | LengthPrefixedString,
+  clarityVersion?: ClarityVersion
+): SmartContractPayload | VersionedSmartContractPayload {
   if (typeof contractName === 'string') {
     contractName = createLPString(contractName);
   }
@@ -113,6 +141,15 @@ export function createSmartContractPayload(
     codeBody = codeBodyString(codeBody);
   }
 
+  if (typeof clarityVersion === 'number') {
+    return {
+      type: StacksMessageType.Payload,
+      payloadType: PayloadType.VersionedSmartContract,
+      clarityVersion,
+      contractName,
+      codeBody,
+    };
+  }
   return {
     type: StacksMessageType.Payload,
     payloadType: PayloadType.SmartContract,
@@ -136,14 +173,33 @@ export interface CoinbasePayload {
   readonly coinbaseBytes: Uint8Array;
 }
 
-export function createCoinbasePayload(coinbaseBytes: Uint8Array): CoinbasePayload {
+export interface CoinbasePayloadToAltRecipient {
+  readonly type: StacksMessageType.Payload;
+  readonly payloadType: PayloadType.CoinbaseToAltRecipient;
+  readonly coinbaseBytes: Uint8Array;
+  readonly recipient: PrincipalCV;
+}
+
+export function createCoinbasePayload(
+  coinbaseBytes: Uint8Array,
+  altRecipient?: PrincipalCV
+): CoinbasePayload | CoinbasePayloadToAltRecipient {
   if (coinbaseBytes.byteLength != COINBASE_BYTES_LENGTH) {
-    throw Error(`Coinbase length must be ${COINBASE_BYTES_LENGTH} bytes`);
+    throw Error(`Coinbase buffer size must be ${COINBASE_BYTES_LENGTH} bytes`);
+  }
+
+  if (altRecipient != undefined) {
+    return {
+      type: StacksMessageType.Payload,
+      payloadType: PayloadType.CoinbaseToAltRecipient,
+      coinbaseBytes,
+      recipient: altRecipient,
+    };
   }
   return {
     type: StacksMessageType.Payload,
     payloadType: PayloadType.Coinbase,
-    coinbaseBytes: coinbaseBytes,
+    coinbaseBytes,
   };
 }
 
@@ -172,11 +228,20 @@ export function serializePayload(payload: PayloadInput): Uint8Array {
       bytesArray.push(serializeStacksMessage(payload.contractName));
       bytesArray.push(serializeStacksMessage(payload.codeBody));
       break;
+    case PayloadType.VersionedSmartContract:
+      bytesArray.push(payload.clarityVersion);
+      bytesArray.push(serializeStacksMessage(payload.contractName));
+      bytesArray.push(serializeStacksMessage(payload.codeBody));
+      break;
     case PayloadType.PoisonMicroblock:
       // TODO: implement
       break;
     case PayloadType.Coinbase:
       bytesArray.push(payload.coinbaseBytes);
+      break;
+    case PayloadType.CoinbaseToAltRecipient:
+      bytesArray.push(payload.coinbaseBytes);
+      bytesArray.push(serializeCV(payload.recipient));
       break;
   }
 
@@ -214,11 +279,24 @@ export function deserializePayload(bytesReader: BytesReader): Payload {
       const smartContractName = deserializeLPString(bytesReader);
       const codeBody = deserializeLPString(bytesReader, 4, 100_000);
       return createSmartContractPayload(smartContractName, codeBody);
+
+    case PayloadType.VersionedSmartContract: {
+      const clarityVersion = bytesReader.readUInt8Enum(ClarityVersion, n => {
+        throw new Error(`Cannot recognize ClarityVersion: ${n}`);
+      });
+      const smartContractName = deserializeLPString(bytesReader);
+      const codeBody = deserializeLPString(bytesReader, 4, 100_000);
+      return createSmartContractPayload(smartContractName, codeBody, clarityVersion);
+    }
     case PayloadType.PoisonMicroblock:
       // TODO: implement
       return createPoisonPayload();
     case PayloadType.Coinbase:
       const coinbaseBytes = bytesReader.readBytes(COINBASE_BYTES_LENGTH);
       return createCoinbasePayload(coinbaseBytes);
+    case PayloadType.CoinbaseToAltRecipient:
+      const coinbaseToAltRecipientBuffer = bytesReader.readBytes(COINBASE_BYTES_LENGTH);
+      const altRecipient = deserializeCV(bytesReader) as PrincipalCV;
+      return createCoinbasePayload(coinbaseToAltRecipientBuffer, altRecipient);
   }
 }