+++ packages/transactions/src/transaction.ts
@@ -83,12 +83,14 @@ export class StacksTransaction {
     } else {
       switch (payload.payloadType) {
         case PayloadType.Coinbase:
+        case PayloadType.CoinbaseToAltRecipient:
         case PayloadType.PoisonMicroblock: {
           this.anchorMode = AnchorMode.OnChainOnly;
           break;
         }
         case PayloadType.ContractCall:
         case PayloadType.SmartContract:
+        case PayloadType.VersionedSmartContract:
         case PayloadType.TokenTransfer: {
           this.anchorMode = AnchorMode.Any;
           break;