+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,40 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.0.2](https://github.com/hirosystems/stacks.js/compare/v6.0.1...v6.0.2) (2022-11-30)
+
+**Note:** Version bump only for package @stacks/cli
+
+
+
+
+
+## [6.0.1](https://github.com/hirosystems/stacks.js/compare/v6.0.0...v6.0.1) (2022-11-24)
+
+**Note:** Version bump only for package @stacks/cli
+
+
+
+
+
+## [6.0.0](https://github.com/hirosystems/stacks.js/compare/v5.0.3...v6.0.0) (2022-11-23)
+
+**Note:** Version bump only for package @stacks/cli
+
+
+
+
+
+## [5.0.3](https://github.com/hirosystems/stacks.js/compare/v5.0.2...v5.0.3) (2022-11-18)
+
+
+### Bug Fixes
+
+* add public key to make_keychain ([2b69958](https://github.com/hirosystems/stacks.js/commit/2b69958b4807105dffb01c61d7ce39cf569437b0))
+* use correct wif encoding for testnet ([a89cc05](https://github.com/hirosystems/stacks.js/commit/a89cc054e7cc17988133364fd8a0ef776ee5ea4f))
+
+
+
 ## [5.0.2](https://github.com/hirosystems/stacks.js/compare/v5.0.1...v5.0.2) (2022-10-19)
 
 **Note:** Version bump only for package @stacks/cli